% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmeU-package.R
\docType{package}
\name{nlmeU-package}
\alias{nlmeU}
\alias{nlmeU-package}
\title{nlmeU: Datasets and Utility Functions Enhancing Functionality of 'nlme' Package}
\description{
Provides functions and datasets to support the book by Galecki and Burzykowski (2013), "Linear Mixed-Effects Models Using R: A Step-by-Step Approach", Springer. Version `r packageVersion("nlmeU")`  includes functions for power calculations, log-likelihood contributions, and data simulation for linear mixed-effects models.
}
\details{
This package provides datasets and utility functions to complement the \code{nlme} package,
including functions like \code{\link{logLik1}}, \code{\link{Pwr}}, and \code{\link{simulateY}}.
It also includes datasets such as \code{\link{armd}}, \code{\link{armd0}}, \code{\link{armd.wide}},
\code{\link{fcat}}, \code{\link{prt.fiber}}, \code{\link{prt}}, \code{\link{prt.subjects}}, and \code{\link{SIIdata}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agalecki/nlmeU}
}

}
\author{
\strong{Maintainer}: Andrzej Galecki \email{agalecki@umich.edu} (\href{https://orcid.org/0000-0003-3058-1312}{ORCID})

Authors:
\itemize{
  \item Tomasz Burzykowski \email{tomasz.burzykowski@uhasselt.be} (\href{https://orcid.org/0000-0003-3378-975X}{ORCID})
}

}
