% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pwr.R
\name{print.Pwr}
\alias{print.Pwr}
\title{Print method for Pwr objects}
\usage{
\method{print}{Pwr}(x, verbose = attr(x, "verbose"), ...)
}
\arguments{
\item{x}{an object of class \code{Pwr.lme}.}

\item{verbose}{logical. If \code{TRUE}, additional information is printed. By default, uses the \code{verbose} attribute of \code{x}.}

\item{\dots}{additional arguments passed to the method.}
}
\value{
Prints the power calculation results and returns \code{NULL} invisibly.
}
\description{
Prints the results of power calculations for objects of class \code{Pwr.lme}.
}
\author{
Andrzej Galecki and Tomasz Burzykowski
}
