% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{runScript}
\alias{runScript}
\title{Execute scripts from Galecki and Burzykowski (2013)}
\usage{
runScript(
  script = NA,
  package = "nlmeU",
  subdir = "scriptsR4.5.1",
  echo = TRUE
)
}
\arguments{
\item{script}{character string containing the name of the script to be executed. By default, \code{NA}.}

\item{package}{character string containing the package name. By default, \code{"nlmeU"}.}

\item{subdir}{subdirectory containing scripts. By default, \code{"scriptsR2.15.0"}.}

\item{echo}{logical. If \code{TRUE}, the script is executed with output printed. Used by \code{source()}. By default, \code{TRUE}.}
}
\value{
The script is executed, and results are printed. If \code{script} is \code{NA}, a list of available scripts is printed.
}
\description{
Executes scripts from the book by Galecki and Burzykowski (2013). If called without arguments, it prints a list of available scripts.
}
\examples{
runScript()
}
\references{
Galecki, A., & Burzykowski, T. (2013). *Linear Mixed-Effects Models: A Step-by-Step Approach*. Springer.
}
\author{
Andrzej Galecki and Tomasz Burzykowski
}
