% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{sigma.default}
\alias{sigma.default}
\title{Default method for extracting scale parameter sigma}
\usage{
\method{sigma}{default}(object, ...)
}
\arguments{
\item{object}{an object for which the scale parameter can be extracted.}

\item{\dots}{additional arguments passed to the method.}
}
\value{
Numeric scalar value representing the scale parameter.
}
\description{
This is a method for the \code{\link{sigma}} generic function.
}
\author{
Andrzej Galecki and Tomasz Burzykowski
}
