% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{sigma}
\alias{sigma}
\title{Extract scale parameter sigma from a model fit}
\usage{
sigma(object, ...)
}
\arguments{
\item{object}{an object for which the scale parameter can be extracted.}

\item{\dots}{some methods for this generic function may require additional arguments.}
}
\value{
Numeric scalar value representing the scale parameter.
}
\description{
This function is generic; method functions can be written to handle specific classes of objects.
}
\examples{
\dontrun{
  library(nlme)
  fm1 <- lme(distance ~ age, data = Orthodont)
  sigma(fm1)
}
}
\author{
Andrzej Galecki and Tomasz Burzykowski
}
