% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateY.R
\name{sigmaTolme}
\alias{sigmaTolme}
\title{Adjusts the scale parameter of an \code{lme} object}
\usage{
sigmaTolme(object, value)
}
\arguments{
\item{object}{an \code{\link[nlme:lme]{lme}} object.}

\item{value}{numeric scalar specifying the new scale parameter value.}
}
\value{
The modified \code{\link[nlme:lme]{lme}} object with rescaled \code{sigma} and variance components.
}
\description{
Adjusts the scale parameter (\code{sigma}) of an \code{\link[nlme:lme]{lme}} object
and rescales variance-related components accordingly. This function is intended
for use with \code{\link{Pwr.lme}}. **Warning**: This function modifies the
\code{lme} object and should be used cautiously to avoid corrupting the model.
}
\author{
Andrzej Galecki and Tomasz Burzykowski
}
\keyword{internal}
