% nlmeU/man/armd0.Rd
% Part of the nlmeU package for R 
% Distributed under GPL 2 or later: see nlmeU/LICENCE
\name{armd0}
\alias{armd0}
\docType{data}
\title{armd0 Data (1107 x 8)}% Data Title
\description{Data armd0
  ...
}
\usage{data(armd0)}
\format{
The \code{armd0} data frame has 1107 rows and 8 columns
\describe{
   \item{subject}{
     a  factor with 240 levels  \code{1}, \code{2}, \code{3}, \code{4}, \code{5}, ...
   }
   \item{treat.f}{
     a  factor with 2 levels  \code{Placebo}, \code{Active}
   }
   \item{visual0}{
     an integer vector with values from 20 to 85
   }
   \item{miss.pat}{
     a  factor with 9 levels  \code{----}, \code{---X}, \code{--X-}, \code{--XX}, \code{-XX-}, ...
   }
   \item{time.f}{
     a  factor with 5 levels  \code{Baseline}, \code{4wks}, \code{12wks}, \code{24wks}, \code{52wks}
   }
   \item{time}{
     a numeric vector with values from 0 to 52
   }
   \item{visual}{
     an integer vector with values from 3 to 85
   }
   \item{tp}{
     a numeric vector with values from 0 to 4
   }
}}
\details{
 ...
}
\source{
  ???? Doe, A. B. and Doe, J. M. (xxxx), \emph{Title},
  Springer, New York.  (Appendix A.17)
}
\examples{
summary(armd0)
}
\keyword{datasets}
