% nlmeU/man/fcat.Rd
% Part of the nlmeU package for R 
% Distributed under GPL 2 or later: see nlmeU/LICENCE
\name{fcat}
\alias{fcat}
\docType{data}
\title{fcat Data (4851 x 3)}% Data Title
\description{Data fcat
  ...
}
\usage{data(fcat)}
\format{
The \code{fcat} data frame has 4851 rows and 3 columns
\describe{
   \item{target}{
     a  factor with 9 levels  \code{T1(4)}, \code{T2(6)}, \code{T3(8)}, \code{T4(5)}, \code{T5(9)}, ...
   }
   \item{id}{
     a  factor with 539 levels  \code{1}, \code{2}, \code{3}, \code{4}, \code{5}, ...
   }
   \item{scorec}{
     an integer vector with values from 0 to 9
   }
}}
\details{
 ...
}
\source{
  ???? Doe, A. B. and Doe, J. M. (xxxx), \emph{Title},
  Springer, New York.  (Appendix A.17)
}
\examples{
summary(fcat)
}
\keyword{datasets}
