\docType{package}
\name{nlmeU-package}
\alias{nlmeU}
\alias{nlmeU-package}
\title{Datasets and auxiliary functions for Galecki and Burzykowski book 2013.}
\description{
  Datasets and auxiliary functions for Galecki and
  Burzykowski book (2013).
}
\details{
  Datasets and auxiliary functions for Galecki and
  Burzykowski book (2013). Package under development.
}
\author{
  \email{agalecki@umich.edu}, \email{Tomasz Burzykowski
  <tomasz.burzykowski@uhasselt.be>}
}
\seealso{
  \code{\link{nlmeUpdK}}
}
\keyword{package}

