\docType{package}
\name{nlmeU-package}
\alias{nlmeU}
\alias{nlmeU-package}
\title{Datasets and auxiliary functions for Galecki and Burzykowski book 2013.}
\description{
  Datasets and auxiliary functions for Galecki and
  Burzykowski book (2013).
}
\details{
  Datasets and auxiliary functions for Galecki and
  Burzykowski book (2013). Package under development.
}
\author{
  Andrzej Galecki \email{agalecki@umich.edu}, Tomasz Burzykowski \email{tomasz.burzykowski@uhasselt.be}
}
\keyword{package}

