% nlmeU/man/prt.Rd
% Part of the nlmeU package for R 
% Distributed under GPL 2 or later: see nlmeU/LICENCE
\name{prt}
\alias{prt}
\docType{data}
\title{prt Data (2471 x 9)}% Data Title
\description{Data prt
  ...
}
\usage{data(prt)}
\format{
The \code{prt} data frame has 2471 rows and 9 columns
\describe{
   \item{id}{
     a  factor with 63 levels  \code{5}, \code{10}, \code{15}, \code{20}, \code{25}, ...
   }
   \item{prt.f}{
     a  factor with 2 levels  \code{High}, \code{Low}
   }
   \item{age.f}{
     a  factor with 2 levels  \code{Young}, \code{Old}
   }
   \item{sex.f}{
     a  factor with 2 levels  \code{Female}, \code{Male}
   }
   \item{bmi}{
     a numeric vector with values from 18.3661027282426 to 32.2863403944485
   }
   \item{iso.fo}{
     a numeric vector with values from 0.16 to 2.565
   }
   \item{spec.fo}{
     a numeric vector with values from 80.5 to 290
   }
   \item{occ.f}{
     a  factor with 2 levels  \code{Pre}, \code{Pos}
   }
   \item{fiber.f}{
     a  factor with 2 levels  \code{Type 1}, \code{Type 2}
   }
}}
\details{
 ...
}
\source{
  ???? Doe, A. B. and Doe, J. M. (xxxx), \emph{Title},
  Springer, New York.  (Appendix A.17)
}
\examples{
summary(prt)
}
\keyword{datasets}
