% File nlmeU/man/Pwr.Rd
% Part of the nlmeU package for R

\name{Pwr}
\title{Calculates power based on a model fit}
\usage{
Pwr(object, ...)
}
\alias{Pwr}
\arguments{
 \item{object}{an object containing the results returned by a model fitting function (e.g., lme).} 
 \item{\dots}{some methods for this generic function may require additional arguments.}
}
\description{
  This function is generic; method functions can be written to handle
  specific classes of objects. 
}
\value{
  numeric scalar value. 
}
\author{Andrzej Galecki and Tomasz Burzykowski}
\seealso{\code{\link{Pwr.lme}}}
%  \code{\link{sigma.pdMat}}}
\examples{
## Pwr (fm1)
}
%\keyword{models}
