% File nlmeU/man/sigma.Rd
% Part of the nlmeU package for R

\name{sigma}
\title{Extract scale parameter sigma from a model fit}
\usage{
sigma(object, ...)
}
\alias{sigma}
\arguments{
 \item{object}{an object for which scale parameter can be
   extracted.} 
 \item{\dots}{some methods for this generic function may require additional
   arguments.}
}
\description{
  This function is generic; method functions can be written to handle
  specific classes of objects. 
}
\value{
  numeric scalar value. 
}
\author{Andrzej Galecki \& Tomasz Burzykowski}
%\seealso{\code{\link{sigma.default}},
%  \code{\link{sigma.pdMat}}}
\examples{
## sigma (fm1)
}
%\keyword{models}
