\name{sigma}
\alias{sigma}
\title{Extract scale parameter sigma from a model fit}
\usage{
  sigma(object, ...)
}
\arguments{
  \item{object}{an object for which scale parameter can be
  extracted.}

  \item{\dots}{some methods for this generic function may
  require additional arguments.}
}
\value{
  numeric scalar value.
}
\description{
  This function is generic; method functions can be written
  to handle specific classes of objects.
}
\examples{
## sigma (fm1)
}
\author{
  Andrzej Galecki and Tomasz Burzykowski
}

